Refactor New structure for pages and ui components and fixed age calculation
