+++ client/src/auth/components/index.ts
@@ -1,3 +1,3 @@
 // Export all authentication components
 export { ProtectedRoute } from "./ProtectedRoute";
-export { default as Login } from "./Login"; 
+export { default as Login } from "@/pages/Login";
