+++ client/src/pages/About.tsx
@@ -1,5 +1,5 @@
 import { Container, Text, Box } from "@chakra-ui/react";
-import Navbar from "./Navbar";
+import Navbar from "@/ui/Navbar";
 
 const About = () => {
   return (
