+++ client/src/pages/index.ts
@@ -1,0 +1,3 @@
+export { default as Login } from './Login';
+export { default as Home } from './Home';
+export { default as About } from './About';
