--- development-workflow-errors.md
@@ -1,146 +1,0 @@
-# Development Workflow Error Report
-
-**Branch:** `fix/lior/cicd-pipeline-v2`  
-**Date:** June 9, 2025  
-**Status:** ❌ FAILED - Multiple check failures
-
-## Summary
-
-The development workflow checks failed with **35 total issues**:
-- **22 ESLint errors**
-- **2 ESLint warnings** 
-- **11 TypeScript build errors**
-- **1 Security audit failure**
-
----
-
-## 1. ESLint Errors (22 errors, 2 warnings)
-
-### TypeScript `any` Type Violations (20 errors)
-**Rule:** `@typescript-eslint/no-explicit-any`
-
-#### Skills Components:
-- `src/features/slideshow-form/components/skills/SelectedSkillsSection.tsx`
-  - Line 11:26 - Unexpected any type
-  - Line 12:27 - Unexpected any type
-
-- `src/features/slideshow-form/components/skills/SkillCategoriesSection.tsx`
-  - Line 9:25 - Unexpected any type
-  - Line 10:27 - Unexpected any type
-
-- `src/features/slideshow-form/components/skills/SkillSubcategoriesSection.tsx`
-  - Line 12:25 - Unexpected any type
-  - Line 13:27 - Unexpected any type
-
-- `src/features/slideshow-form/components/skills/SkillsStep.tsx`
-  - Line 18:25 - Unexpected any type
-  - Line 19:26 - Unexpected any type
-  - Line 20:27 - Unexpected any type
-
-#### Dashboard Cards:
-- `src/features/slideshow-form/dashboard/main-layout/section-cards/ExperienceCard.tsx`
-  - Line 65:28 - Unexpected any type
-  - Line 65:40 - Unexpected any type
-
-- `src/features/slideshow-form/dashboard/main-layout/section-cards/HardSkillsCard.tsx`
-  - Line 32:48 - Unexpected any type
-  - Line 32:60 - Unexpected any type
-
-#### Hooks:
-- `src/features/slideshow-form/hooks/useLoadCvData.ts`
-  - Line 26:23 - Unexpected any type
-  - Line 27:35 - Unexpected any type
-
-- `src/features/slideshow-form/hooks/useSaveCvData.ts`
-  - Line 7:80 - Unexpected any type
-
-#### Steps:
-- `src/features/slideshow-form/steps/StepVitals.tsx`
-  - Line 23:4 - Unexpected any type
-  - Line 128:4 - Unexpected any type
-  - Line 176:51 - Unexpected any type
-  - Line 211:51 - Unexpected any type
-
-#### Pages:
-- `src/pages/Home.tsx`
-  - Line 22:9 - Unexpected any type
-
-### Unused Variables (1 error)
-**Rule:** `@typescript-eslint/no-unused-vars`
-
-- `src/features/slideshow-form/hooks/useHardSkills.ts`
-  - Line 78:17 - Variable '_' is assigned but never used
-
-### React Fast Refresh Warnings (2 warnings)
-**Rule:** `react-refresh/only-export-components`
-
-- `src/components/ui/toaster.tsx`
-  - Line 12:14 - Fast refresh only works when a file only exports components
-
-- `src/context/AuthContext.tsx`
-  - Line 107:14 - Fast refresh only works when a file only exports components
-
----
-
-## 2. TypeScript Build Errors (11 errors)
-
-### ListCollection Type Mismatches (7 errors)
-
-#### Preferences Form Fields:
-- `src/features/slideshow-form/components/preferences/PreferencesFormFields.tsx`
-  - Line 49:9 - `cvPurposeCollection` type mismatch with `CollectionItem[]`
-  - Line 60:9 - `professionalPreferenceCollection` type mismatch with `CollectionItem[]`
-  - Line 71:9 - `experienceLevelCollection` type mismatch with `CollectionItem[]`
-  - Line 84:9 - `industryCollection` type mismatch with `CollectionItem[]`
-  - Line 97:9 - `salaryRangeCollection` type mismatch with `CollectionItem[]`
-
-#### Education Step:
-- `src/features/slideshow-form/steps/StepEducation.tsx`
-  - Line 81:15 - `degreesCollection` type mismatch with `CollectionItem[]`
-  - Line 91:15 - `fieldsOfStudyCollection` type mismatch with `CollectionItem[]`
-
-### String Type Assignment Errors (2 errors)
-
-- `src/features/slideshow-form/components/preferences/PreferencesFormFields.tsx`
-  - Line 52:57 - Cannot assign `string` to `CvPurpose` type
-  - Line 63:70 - Cannot assign `string` to `ProfessionalPreference` type
-
-### ComboboxField Type Error (2 errors)
-
-- `src/features/slideshow-form/steps/StepEducation.tsx`
-  - Line 91:15 - Type mismatch in ComboboxField component props
-
----
-
-## 3. Security Audit Failure
-
-**Issue:** npm audit failed due to network/TLS configuration
-**Error:** `426 Upgrade Required - POST http://registry.npmjs.org`
-**Cause:** npm registry requires HTTPS and TLS 1.2+ connections
-
----
-
-## Priority Fixes Required
-
-### High Priority (Blocks Build)
-1. **Fix ListCollection type issues** - Update type definitions for collection props
-2. **Resolve string type assignments** - Proper type casting for enum values
-3. **Fix ComboboxField type errors** - Update component prop types
-
-### Medium Priority (Code Quality)
-1. **Replace all `any` types** - Define proper TypeScript interfaces
-2. **Remove unused variables** - Clean up unused imports/variables
-
-### Low Priority (Warnings)
-1. **Split React components** - Separate constants/functions from component exports
-2. **Update npm registry configuration** - Enable HTTPS for security audits
-
----
-
-## Next Steps
-
-1. Address TypeScript build errors first (prevents compilation)
-2. Fix ESLint `any` type violations for better type safety
-3. Clean up unused variables and warnings
-4. Configure npm for HTTPS registry access
-5. Re-run development workflow checks to verify fixes
