+++ package.json
@@ -1,0 +1,34 @@
+{
+    "name": "career-vision",
+    "version": "1.0.0",
+    "description": "A comprehensive digital platform empowering job seekers in the Israeli tech industry with AI-powered CV generation, interview coaching, and career guidance tools",
+    "private": true,
+    "scripts": {
+        "dev": "concurrently \"npm run server\" \"npm run client\"",
+        "server": "cd server && npm run dev",
+        "client": "cd client && npm run dev",
+        "install:all": "npm install && cd server && npm install && cd ../client && npm install",
+        "build": "cd client && npm run build",
+        "start": "cd server && npm start",
+        "clean": "rm -rf node_modules server/node_modules client/node_modules",
+        "lint": "concurrently \"cd server && npm run lint\" \"cd client && npm run lint\"",
+        "test": "concurrently \"cd server && npm test\" \"cd client && npm test\""
+    },
+    "workspaces": [
+        "server",
+        "client"
+    ],
+    "engines": {
+        "node": ">=18.0.0",
+        "npm": ">=8.0.0"
+    },
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/TCP-TechCyberPoint/CV-CareerVision"
+    },
+    "author": "Career Vision <careervision2025@gmail.com>",
+    "license": "MIT",
+    "devDependencies": {
+        "concurrently": "^9.2.0"
+    }
+}
